requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.0.1
selenium==4.29.0
webdriver-manager==4.0.2
yt-dlp==2025.2.19